+++ vcmi_gym/envs/v0/util/pyconnector.py
@@ -39,7 +39,7 @@ ERRSIZE = len(ERRMAP)
 ERRNAMES = [errname for (errname, _) in ERRMAP.values()]
 ERRFLAGS = list(ERRMAP.keys())
 
-PyState = ctypes.c_int * STATE_SIZE
+PyState = ctypes.c_float * STATE_SIZE
 PyAction = ctypes.c_int
 PyActmask = ctypes.c_bool * N_ACTIONS
